site: add sample index.html with search bar
this is a rough draft of a sample landing page with a search bar and
nothing else.  the page relies on jquery.

the search bar autocompletes the following cities:
 - San Francisco
 - New York
 - London
 - Amsterdam
 - Venice
 - St. Petersburg
 - Shanghai
 - Melbourne

the drop down menu to the right is not functional, yet.  in theory, the
drop down menu would allow some sort of search filtering based on
country/region, continent, or some other criteria.