display search results via while loop
the use of a while loop without a counter to display search results will
iterate through all of the results, no matter how many there are.  in a
future implementation, consider using a limited for loop or breaking
from the while loop after a counter reaches a certain value, the value
which is ideally selected from something like a drop-down `display n
results' menu.

also complete:
 - remove branding